# draw-and-guess-server/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for Draw and Guess multiplayer game",
        "title": "Draw and Guess Game API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/app",
    "paths": {
        "/game/room": {
            "post": {
                "description": "Create a new game room with the specified drawer",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "game-rooms"
                ],
                "summary": "Create a new game room",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Username of the room creator",
                        "name": "ldap_user",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.ApiResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/models.GameRoom"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ApiResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ApiResult"
                        }
                    }
                }
            }
        },
        "/game/room/{id}": {
            "delete": {
                "description": "Delete a game room (host only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "game-rooms"
                ],
                "summary": "Delete game room",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Room ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ApiResult"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ApiResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ApiResult"
                        }
                    }
                }
            },
            "patch": {
                "description": "Update game room properties or advance to the next round",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "game-rooms"
                ],
                "summary": "Update game room",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Room ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Action to perform (e.g., 'advance_round')",
                        "name": "action",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "description": "Number of users",
                        "name": "user_count",
                        "in": "formData"
                    },
                    {
                        "type": "boolean",
                        "description": "Room active status",
                        "name": "is_active",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.ApiResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/models.GameRoom"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ApiResult"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ApiResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ApiResult"
                        }
                    }
                }
            }
        },
        "/game/room/{id}/answer": {
            "post": {
                "description": "Check if the submitted answer is correct",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "game-rooms"
                ],
                "summary": "Check answer",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Room ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Username",
                        "name": "ldap_user",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Answer to check",
                        "name": "answer",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.ApiResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/models.GameRoom"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ApiResult"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ApiResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ApiResult"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.ApiResult": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "result": {},
                "status": {
                    "type": "boolean"
                }
            }
        },
        "models.GameRoom": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "current_word": {
                    "type": "string"
                },
                "drawer_user": {
                    "description": "방장 (그림 그리는 사람)",
                    "type": "string"
                },
                "game_status": {
                    "description": "waiting, playing, finished",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "max_rounds": {
                    "description": "최대 라운드 (3)",
                    "type": "integer"
                },
                "players": {
                    "description": "참가자 목록",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Player"
                    }
                },
                "round_number": {
                    "description": "현재 라운드 (1-3)",
                    "type": "integer"
                },
                "time_left": {
                    "description": "남은 시간 (초)",
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "used_words": {
                    "description": "사용된 단어들",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "uuid": {
                    "type": "string"
                },
                "winning_score": {
                    "description": "우승 점수 (3)",
                    "type": "integer"
                }
            }
        },
        "models.Player": {
            "type": "object",
            "properties": {
                "attempts": {
                    "description": "이번 라운드 제출 횟수",
                    "type": "integer"
                },
                "score": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    }
}